Add files via upload
This is the version of the data that were used for the analyses in the manuscript and range from 1995 to 2014. These data should be used for pure replication in conjunction with the R Script and Stata .Do file. The extended data from 1995 to 2016 should not be used for pure replication.